requirements.txt
numpy
pillow
scipy
argparse
